+++ spec/features/project_groups/users_can_view_project_groups_spec.rb
@@ -5,13 +5,16 @@ RSpec.feature "Users can view project groups" do
   let(:project) { create(:conversion_project, caseworker: user) }
 
   before do
-    mock_successful_api_responses(urn: 123456, ukprn: 10061021)
+    mock_all_academies_api_responses
     sign_in_with_user(user)
+
+    create(:project_group, group_identifier: "GRP_12345678")
   end
 
   scenario "they can view all groups" do
     click_on "Groups"
 
     expect(page).to have_content "Groups"
+    expect(page).to have_content "GRP_12345678"
   end
 end